fixed some issues in the MapDatabase with the new map file format and refactored some code.
git-svn-id: http://mapsforge.googlecode.com/svn/trunk@764 a424f762-2dd8-11df-bfeb-7b4c664f286a

+++ mapsforge/src/org/mapsforge/android/maps/DatabaseMapGenerator.java
@@ -2052,7 +2052,7 @@ abstract class DatabaseMapGenerator extends MapGenerator implements
 	 */
 	final void renderWay(byte wayLayer, byte wayNumberOfRealTags, String wayName,
 			String wayRef, int[] wayLabelPosition, boolean[] wayTagIds, byte wayTagBitmap,
-			short wayNodesSequenceLength, int[] wayNodesSequence, int[][] innerWays) {
+			int wayNodesSequenceLength, int[] wayNodesSequence, int[][] innerWays) {
 		this.remainingTags = wayNumberOfRealTags;
 		if (innerWays == null) {
 			this.coordinates = new float[1][];
